Move tags.js to new "data" folder, move addonGroups there
